Airline Reservation System (Project Code and Requirements)
This programming project involves developing a Airline Reservation
System for use by
travel professionals (i.e. not passengers). It is a Java-based GUI
application that
interfaces with a backend MySQL database to manage its data. The GUI is
browser based (JSP)
The database schema is based on Figure 3.8 (page 82) in the textbook. A
description of the high-level functionality can be found in Exercise
3.12 (page 81). An
ER diagram can be found in Figure 7.20 (page 236), which provides a
conceptual model.